fix: address Humanizer migration findings — i64 support, MCP improvements
Compiler/Language fixes:
- Fix integer overflow data corruption: widen IntLiteralNode and
  BoundIntLiteral from int to long, preventing silent truncation of
  C# long literals during C#→Calor conversion
- Lexer: add long.TryParse fallback for values outside int range
- Lexer: support numeric separators (1_000_000) by consuming underscores
- Lexer: fix hex literal truncation — store as long when > int.MaxValue
- CSharpEmitter: add L suffix for values outside int range
- RoslynSyntaxVisitor: remove (int)longVal cast that caused data corruption

MCP improvements:
- McpServer: set Console.InputEncoding/OutputEncoding to UTF-8 in
  CreateStdio() to fix § character corruption on Windows
- BatchConvertTool: return structured validation errors with line,
  column, category, and suggestion fields (backward-compatible: keeps
  errors field alongside new issues field)
- CompileTool: add errorCategories summary to batch compilation output
  showing error counts grouped by error code
- DiagnoseTool: add 5 new error-suggestion patterns for common
  converter output issues (§NEW{object}, integer overflow, generic
  type params, IF ID mismatch)

Tests:
- 10 new lexer tests for long literals, numeric separators, hex, boundaries
- 4 new E2E tests for long literal compilation and L suffix emission
- 1 new converter test verifying long literal preservation

+++ src/Calor.Compiler/Ast/ExpressionNodes.cs
@@ -16,9 +16,9 @@ public abstract class ExpressionNode : AstNode
 /// </summary>
 public sealed class IntLiteralNode : ExpressionNode
 {
-    public int Value { get; }
+    public long Value { get; }
 
-    public IntLiteralNode(TextSpan span, int value) : base(span)
+    public IntLiteralNode(TextSpan span, long value) : base(span)
     {
         Value = value;
     }
